+++ zabbix_utils/aiosender.py
@@ -69,6 +69,7 @@ Defaults to `None`.
         self.use_ipv6 = use_ipv6
         self.tls = {}
 
+        self.host = None
         self.source_ip = None
         self.chunk_size = chunk_size
         self.compression = compression
@@ -105,6 +106,7 @@ Defaults to `None`.
         for cluster in server_row.split(','):
             self.clusters.append(Cluster(cluster.strip().split(';')))
 
+        self.host = config.get('Hostname')
         if 'SourceIP' in config:
             self.source_ip = config.get('SourceIP')
 
@@ -310,4 +312,4 @@ It must be a list of ItemValue objects: {json.dumps(chunk)}")
             TrapperResponse: Response from Zabbix server/proxy.
         """
 
-        return await self.send([ItemValue(host, key, value, clock, ns)])
+        return await self.send([ItemValue(host or self.host or '', key, value, clock, ns)])